Add bson feature support
