+++ cynic/Cargo.toml
@@ -14,7 +14,7 @@ documentation = "https://docs.rs/cynic"
 
 [features]
 default = []
-all = ["chrono"]
+all = ["chrono", "bson"]
 
 [dependencies]
 json-decode = "0.5.0"
@@ -22,6 +22,7 @@ serde = { version = "1.0.104", features = [ "derive" ] }
 serde_json = "1.0"
 cynic-proc-macros = { path = "../cynic-proc-macros", version = "0.9.0" }
 chrono = { version = "0.4.11", optional = true}
+bson = { version = "1.1.0", optional = true}
 
 [dev-dependencies]
 maplit = "1.0.2"
